You can now throw 5 stones!
StoneThrowController updated so you can now throw 5 stones in a row (!). The Unity Inspector is weird with lists, so I put a manual list of 5 stones references (which you then drag and drop in the inspector to load in specific stones) in the StoneThrowController script. We'll probably want to sort that out at some point.